@@ -1,6 +1,6 @@
 {
   "name": "react-schema-form-example",
-  "version": "0.3.8",
+  "version": "0.4.1",
   "description": "React json schema form example",
   "keywords": [
     "json schema",
@@ -24,9 +24,6 @@
     "url": "https://github.com/networknt/react-schema-form.git"
   },
   "dependencies": {
-    "lodash": "^4.16.6",
-    "material-ui": "^0.18.7",
-    "react": "^15.3.2",
     "babel-cli": "^6.24.1",
     "babel-core": "^6.5.2",
     "babel-eslint": "^7.1.0",
@@ -42,13 +39,14 @@
     "jquery": "^3.1.1",
     "less": "^2.7.1",
     "less-loader": "^2.2.3",
-    "react-ace": "4.0.0",
-    "react-dom": "^15.3.2",
+    "lodash": "^4.17.10",
+    "material-ui": "^1.0.0-beta.45",
+    "react": "^16.3.2",
+    "react-ace": "^6.1.0",
+    "react-dom": "^16.3.2",
     "react-hot-loader": "^3.0.0-beta.6",
-    "react-schema-form-rc-select": "^0.2.1",
     "react-schema-form": "file:../",
     "react-select": "^1.0.0-rc.5",
-    "react-tap-event-plugin": "^2.0.0",
     "style-loader": "^0.13.1",
     "webpack": "^2.7.0",
     "webpack-dev-server": "^2.6.1"
