@@ -25,12 +25,10 @@
   },
   "dependencies": {
     "classnames": "^2.2.5",
-    "lodash": "^4.16.6",
     "notevil": "^1.1.0",
+    "lodash": "^4.17.10",
     "objectpath": "^1.2.1",
-    "react-autosuggest": "^9.3.4",
-    "tv4": "^1.2.7",
-    "webpack-config": "^7.5.0"
+    "tv4": "^1.3.0"
   },
   "devDependencies": {
     "babel-cli": "^6.24.1",
@@ -48,26 +46,23 @@
     "jest-cli": "^16.0.2",
     "less": "^2.7.1",
     "less-loader": "^2.2.3",
-    "material-ui": "^0.18.7",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-addons-test-utils": "^15.3.2",
+    "material-ui": "^1.0.0-beta.45",
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2",
     "react-hot-loader": "^3.0.0-beta.6",
-    "react-tap-event-plugin": "^2.0.0",
     "rimraf": "^2.5.4",
     "style-loader": "^0.13.1",
     "webpack": "^2.7.0"
   },
   "peerDependencies": {
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "material-ui": "^0.18.7"
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2"
   },
   "scripts": {
     "lint": "eslint src",
     "preversion": "webpack -p && git add dist",
-    "prepublish": "rimraf lib && babel src -d lib",
-    "test": "jest"
+    "prepublishOnly": "rimraf lib && babel src -d lib",
+    "test": "npm run lint && jest"
   },
   "jest": {
     "testPathDirs": [
