@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-
-var ExamplePage  = require('./ExamplePage');
+import ExamplePage  from './ExamplePage';
 
 // For me this obfuscates the location of the error when searching from console. Not sure why it's necessary.
 var warn = console.warn;
@@ -15,7 +14,7 @@ console.warn = function(warning) {
 //Can go away when react 1.0 release
 //Check this repo:
 //https://github.com/zilverline/react-tap-event-plugin
-import injectTapEventPlugin from 'react-tap-event-plugin';
-injectTapEventPlugin();
+//import injectTapEventPlugin from 'react-tap-event-plugin';
+//injectTapEventPlugin();
 
 ReactDOM.render(<ExamplePage />, document.getElementById("app"));
