@@ -14,53 +14,17 @@ var _utils = require('./utils');
 
 var _utils2 = _interopRequireDefault(_utils);
 
-var _Number = require('./Number');
-
-var _Number2 = _interopRequireDefault(_Number);
-
-var _Text = require('./Text');
-
-var _Text2 = _interopRequireDefault(_Text);
-
-var _TextArea = require('./TextArea');
-
-var _TextArea2 = _interopRequireDefault(_TextArea);
-
-var _Select = require('./Select');
-
-var _Select2 = _interopRequireDefault(_Select);
-
-var _Radios = require('./Radios');
-
-var _Radios2 = _interopRequireDefault(_Radios);
-
-var _Date = require('./Date');
-
-var _Date2 = _interopRequireDefault(_Date);
-
-var _Checkbox = require('./Checkbox');
-
-var _Checkbox2 = _interopRequireDefault(_Checkbox);
-
-var _Help = require('./Help');
-
-var _Help2 = _interopRequireDefault(_Help);
-
 var _ComposedComponent = require('./ComposedComponent');
 
 var _ComposedComponent2 = _interopRequireDefault(_ComposedComponent);
 
-var _RaisedButton = require('material-ui/RaisedButton');
+var _Button = require('material-ui/Button');
 
-var _RaisedButton2 = _interopRequireDefault(_RaisedButton);
+var _Button2 = _interopRequireDefault(_Button);
 
 var _lodash = require('lodash');
 
 var _lodash2 = _interopRequireDefault(_lodash);
-
-var _SchemaForm = require('./SchemaForm');
-
-var _SchemaForm2 = _interopRequireDefault(_SchemaForm);
 
 var _IconButton = require('material-ui/IconButton');
 
@@ -230,7 +194,7 @@ var Array = function (_React$Component) {
                         arrays
                     )
                 ),
-                _react2.default.createElement(_RaisedButton2.default, { label: this.props.form.add || 'Add', secondary: true, onTouchTap: this.onAppend })
+                _react2.default.createElement(_Button2.default, { variant: 'raised', label: this.props.form.add || 'Add', secondary: true, onTouchTap: this.onAppend })
             );
         }
     }]);
