@@ -10,25 +10,13 @@ var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
-var _utils = require('./utils');
-
-var _utils2 = _interopRequireDefault(_utils);
-
-var _classnames = require('classnames');
-
-var _classnames2 = _interopRequireDefault(_classnames);
-
 var _ComposedComponent = require('./ComposedComponent');
 
 var _ComposedComponent2 = _interopRequireDefault(_ComposedComponent);
 
-var _RadioButton = require('material-ui/RadioButton');
+var _Radio = require('material-ui/Radio');
 
-var _RadioButton2 = _interopRequireDefault(_RadioButton);
-
-var _RadioButtonGroup = require('material-ui/RadioButton/RadioButtonGroup');
-
-var _RadioButtonGroup2 = _interopRequireDefault(_RadioButtonGroup);
+var _Radio2 = _interopRequireDefault(_Radio);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -51,7 +39,7 @@ var Radios = function (_React$Component) {
         key: 'render',
         value: function render() {
             var items = this.props.form.titleMap.map(function (item, index) {
-                return _react2.default.createElement(_RadioButton2.default, { label: item.name,
+                return _react2.default.createElement(_Radio2.default, { label: item.name,
                     value: item.value,
                     key: index,
                     disabled: this.props.form.readonly
@@ -67,7 +55,7 @@ var Radios = function (_React$Component) {
                     this.props.form.title
                 ),
                 _react2.default.createElement(
-                    _RadioButtonGroup2.default,
+                    _Radio.RadioGroup,
                     { defaultSelected: this.props.value, name: this.props.form.title, onChange: this.props.onChangeValidate },
                     items
                 )
