@@ -14,13 +14,11 @@ var _ComposedComponent = require('./ComposedComponent');
 
 var _ComposedComponent2 = _interopRequireDefault(_ComposedComponent);
 
-var _MenuItem = require('material-ui/MenuItem');
+var _Menu = require('material-ui/Menu');
 
-var _MenuItem2 = _interopRequireDefault(_MenuItem);
+var _Select = require('material-ui/Select');
 
-var _SelectField = require('material-ui/SelectField');
-
-var _SelectField2 = _interopRequireDefault(_SelectField);
+var _Select2 = _interopRequireDefault(_Select);
 
 var _lodash = require('lodash');
 
@@ -98,19 +96,21 @@ var Select = function (_React$Component) {
         key: 'render',
         value: function render() {
             var menuItems = this.props.form.titleMap.map(function (item, idx) {
-                return _react2.default.createElement(_MenuItem2.default, { key: idx,
-                    primaryText: item.name,
-                    value: item.value });
+                return _react2.default.createElement(
+                    _Menu.MenuItem,
+                    { key: idx, value: item.value },
+                    item.name
+                );
             });
 
             return _react2.default.createElement(
                 'div',
                 { className: this.props.form.htmlClass },
                 _react2.default.createElement(
-                    _SelectField2.default,
+                    _Select2.default,
                     {
                         value: this.state.currentValue,
-                        floatingLabelText: this.props.form.title,
+                        placeholder: this.props.form.title,
                         disabled: this.props.form.readonly,
                         onChange: this.onSelected,
                         fullWidth: true },
