@@ -50,23 +50,19 @@ var Text = function (_React$Component) {
     _createClass(Text, [{
         key: 'render',
         value: function render() {
-            var _props = this.props,
-                form = _props.form,
-                error = _props.error,
-                value = _props.value,
-                onChangeValidate = _props.onChangeValidate;
-
-            return _react2.default.createElement(_TextField2.default, {
-                type: form.type,
-                label: form.title,
-                placeholder: form.placeholder,
-                helperText: error || form.description,
-                error: !!error,
-                onChange: onChangeValidate,
-                value: value,
-                disabled: form.readonly,
-                fullWidth: true
-            });
+            //console.log('Text props', this.props.form.readonly);
+            return _react2.default.createElement(
+                'div',
+                { className: this.props.form.htmlClass },
+                _react2.default.createElement(_TextField2.default, {
+                    type: this.props.form.type,
+                    label: this.props.form.title,
+                    helperText: this.props.form.placeholder,
+                    onChange: this.props.onChangeValidate,
+                    defaultValue: this.props.value,
+                    disabled: this.props.form.readonly
+                })
+            );
         }
     }]);
 
