@@ -3,18 +3,9 @@
  */
 import React from 'react';
 import utils from './utils';
-import Number from './Number';
-import Text from './Text';
-import TextArea from './TextArea';
-import Select from './Select';
-import Radios from './Radios';
-import Date from './Date';
-import Checkbox from './Checkbox';
-import Help from './Help';
 import ComposedComponent from './ComposedComponent';
-import RaisedButton from 'material-ui/RaisedButton';
+import Button from 'material-ui/Button';
 import _ from 'lodash';
-import SchemaForm from './SchemaForm';
 import IconButton from 'material-ui/IconButton';
 
 class Array extends React.Component {
@@ -148,7 +139,7 @@ class Array extends React.Component {
                         {arrays}
                     </ol>
                 </div>
-                <RaisedButton label={this.props.form.add || 'Add'} secondary onTouchTap={this.onAppend}/>
+                <Button variant="raised" label={this.props.form.add || 'Add'} secondary={true} onTouchTap={this.onAppend}/>
             </div>
         );
     }
