@@ -1,20 +1,16 @@
 import React from 'react';
-import utils from './utils';
-import classNames from 'classnames';
 import ComposedComponent from './ComposedComponent';
-
-import RadioButton from 'material-ui/RadioButton';
-import RadioButtonGroup from 'material-ui/RadioButton/RadioButtonGroup';
+import Radio, { RadioGroup } from 'material-ui/Radio';
 
 class Radios extends React.Component {
 
     render() {
         let items = this.props.form.titleMap.map(function(item, index) {
             return (
-                <RadioButton label={item.name}
-                             value={item.value}
-                             key={index}
-                             disabled={this.props.form.readonly}
+                <Radio label={item.name}
+                       value={item.value}
+                       key={index}
+                       disabled={this.props.form.readonly}
                     />
             )
         }.bind(this));
@@ -22,9 +18,9 @@ class Radios extends React.Component {
         return (
             <span className={this.props.form.htmlClass}>
               <label className="control-lable">{this.props.form.title}</label>
-              <RadioButtonGroup defaultSelected={this.props.value} name={this.props.form.title} onChange={this.props.onChangeValidate}>
+              <RadioGroup defaultSelected={this.props.value} name={this.props.form.title} onChange={this.props.onChangeValidate}>
                   {items}
-              </RadioButtonGroup>
+              </RadioGroup>
             </span>
         );
     }
