@@ -1,6 +1,3 @@
-/**
- * Created by steve on 15/09/15.
- */
 import React from 'react';
 import ComposedComponent from './ComposedComponent';
 import TextField from 'material-ui/TextField';
@@ -22,13 +19,12 @@ class Text extends React.Component {
             <div className={this.props.form.htmlClass}>
                 <TextField
                     type={this.props.form.type}
-                    floatingLabelText={this.props.form.title}
-                    hintText={this.props.form.placeholder}
-                    errorText={this.props.error || this.props.errorText}
+                    label={this.props.form.title}
+                    helperText={this.props.form.placeholder}
                     onChange={this.props.onChangeValidate}
                     defaultValue={this.props.value}
                     disabled={this.props.form.readonly}
-                    style={this.props.form.style || {width: '100%'}} />
+                 />
             </div>
         );
     }
